requirements.txt
openpyxl>=3.1.0
requests>=2.31.0
msal>=1.28.0
python-dateutil>=2.9.0
jinja2>=3.1.0
python-dotenv>=1.0.0
